ath11k: Fix QCN9074 firmware boot on x86
2.7.0 series of QCN9074 firmware requests 5 segments of
memory instead of 3, as in the previous series of firmware.
The first segment is too large to be kalloc'd in one go
and requires piecemeal 1MB allocations, but ath11k will
only break the memory requests if there were fewer than 3
segments requested by the firmware.

It seems 5 segments works fine and allows QCN9074 to boot
on x86, so change it accordingly.

Signed-off-by: Tyler J. Stachecki <stachecki.tyler@gmail.com>