wifi: ath11k: Use proper API to get peer id for WCN6750
Currently on WCN6750, we are seeing an issue where signal average field
value in station dump is incorrect. This is because of registering an
incorrect mpdu_info_get_peerid() API for WCN6750. Due to this, we get an
invalid peer id while processing the peer stats, and the signal average
field will have invalid values in station dump.

Fix this by registering correct mpdu_info_get_peerid() API for WCN6750.

Tested-on: WCN6750 hw1.0 AHB WLAN.MSL.1.0.1-00887-QCAMSLSWPLZ-1

Signed-off-by: Youghandhar Chintala <quic_youghand@quicinc.com>