wifi: ath11k: Align RX MPDU info with the hardware for WCN6750
Currently on WCN6750, we are seeing an issue where signal average field
value in station dump is incorrect. This is because RX MPDU info fields
in the driver are not properly aligned with the hardware. Due to this,
we get an invalid peer id while processing the peer stats, and the
signal avg value will not be updated.

Fix this by aligning the RX MPDU info in the driver with the hardware.
Change is applicable for WCN6855 as well.

Tested-on: WCN6750 hw1.0 AHB WLAN.MSL.1.0.1-00887-QCAMSLSWPLZ-1

Signed-off-by: Youghandhar Chintala <quic_youghand@quicinc.com>