HACK: Disable SAM D0 entry/exit notifications
In testing suspend-to-RAM support, it was noticed that the
machine sometimes hangs coming out of S3. The root cause is
SAM changing D0 state.

I believe a device link between the SAM and QUP devices
must be established in order for this to work reliably. It
looks like its only being done for the ACPI path currently.

Signed-off-by: Tyler J. Stachecki <stachecki.tyler@gmail.com>