rtc-pm8xxx: Add qcom_qseecom_uefisecapp dependency
Add a dependency on uefisecapp as the rtc-pm8xxx uses EFI
variables that are protected by the same in order to
function correctly in some cases.

Also move the rtc-pm8xxx initcall to late so that it loads
after uefisecapp registers itself to handle efivarfs access
such that the EFI variable reads/writes work in cases where
access needs to be delegated through it.

Signed-off-by: Tyler J. Stachecki <stachecki.tyler@gmail.com>