drm/msm: Add a ftrace for GPU register access
With IFPC, there is a probability of accessing a GX domain register when
it is collapsed, which leads to GMU fence errors. To debug this, we need
to trace every GPU register accesses and identify the one just before a
GMU fence error. So, add an ftrace to track all GPU register accesses.

Signed-off-by: Akhil P Oommen <akhilpo@quicinc.com>
Signed-off-by: Tyler J. Stachecki <stachecki.tyler@gmail.com>